Set duration detail in playlist to 4 decimals

+++ src/Modules/Video/Entities/Repository/Disk/PlaylistMasterDisk.php
@@ -174,7 +174,7 @@ class PlaylistMasterDisk extends AbstractModule
 
                         $line = trim($line);
                         if (strpos($line, '#EXTINF:') === 0) {
-                            $duration = (float) str_replace('#EXTINF:', '', $line);
+                            $duration = str_replace('#EXTINF:', '', $line);
                         }
                         elseif (strpos($line, '#') !== 0 && $duration > 0.0) {
                             $filename = $line;
@@ -198,7 +198,6 @@ class PlaylistMasterDisk extends AbstractModule
                             $duration = 0.0;
                         }
                     }
-
                     fclose($handle);
                 }
             }
